Refactor: move append_membership to ReplicationHandler
