+++ openraft/src/engine/mod.rs
@@ -40,7 +40,6 @@ mod log_id_list;
 #[cfg(test)] mod initialize_test;
 #[cfg(test)] mod install_snapshot_test;
 #[cfg(test)] mod leader_append_entries_test;
-#[cfg(test)] mod leader_append_membership_test;
 #[cfg(test)] mod log_id_list_test;
 #[cfg(test)] mod startup_test;
 #[cfg(test)] mod testing;
